Add VB_INIT_OUT_ENABLE_ALTERNATE_OS flag
This adds a flag to the list of values returned by VbInit(). When
this flag is set, the BIOS may be asked to boot something other than
ChromeOS. If this requires some sort of special preparation, the BIOS
should do it.

BUG=chromium-os:22454
TEST=none

There is no test for this. It requires a change to the BIOS in order
to do anything differently, and we haven't yet decided whether the
BIOS should pay attention to it.

Reviewed-on: https://gerrit.chromium.org/gerrit/11714
Tested-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>

(cherry picked from commit 0d11efb0dc1d8d2b5eafdd5b65bce82e73fdeecc)

Change-Id: I9507fad6486a3304a45067d8b51a99708e15bbff
Reviewed-on: https://gerrit.chromium.org/gerrit/13258
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Stefan Reinauer <reinauer@chromium.org>