Starting point for futility tool
This is a bare-bones change to start work on the new futility tool. It adds
a new directory and a new test, but doesn't change anything else.

The new test is NOT run by default, since it fails every test case (futility
doesn't do anything useful yet, duh).

BUG=chromium-os:37062
BRANCH=none
TEST=manual

Run

  make vboot_next

This should build things and then report a bunch of test failures.

Change-Id: I5a3336926b684c2a1b0fc6b89976fae327853671
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40353
Reviewed-by: Randall Spangler <rspangler@chromium.org>