vboot: Don't abort for OPROM_MISMATCH if not using RO_NORMAL
When the BIOS calls VbInit() in firmware it may not yet know if
the system is in developer mode if there is a virtual developer mode
switch, instead it relies on the VbNv that is prepared by VbInit().

This means that BIOS may not always pass in the correct value for
OPROM_LOADED because the option rom loading has not actually
happened yet.

However since it has not happened yet the BIOS can still respond to
the VbNv flag being set without needing to reboot, so VbInit() does
not need to abort.  This was causing FAFT tests to get confused.

If using RO_NORMAL then we can assume an old style where VbInit() is
called by the bootloader after Option ROM loading has already happened,
so it does need to be told to reboot in this case.

BUG=chrome-os-partner:32379
BRANCH=samus
TEST=pass FAFT tests on samus

Change-Id: Id3cdfb338a0c7a9901e0c579acce4fc9c7ba016b
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/224631
Reviewed-by: Bill Richardson <wfrichar@chromium.org>