vboot_display: Keep track of displayed screen
Vboot needs to track the currently displayed screen so when it needs to
change the locale or display the debug overlay it knows which screen to
redraw.  Currently only the legacy path is doing this so change the new
path to update the current screen if it is successfully drawn.

BUG=chrome-os-partner:49766
BRANCH=glados
TEST=boot on glados in dev mode, hit tab and ensure screen does not go black

Change-Id: I4a2bf028275db57b2d0469fc1cb574e871820713
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/324549
(cherry picked from commit 00fa2601da5c3ec449dbeca19cc409a6777569fd)
Reviewed-on: https://chromium-review.googlesource.com/325008