detachables: Add unit tests
Initial unit tests for detachable menu UI.  With these tests, the
coverage is at 81.2% for lines for code and 87.5% for function
coverage.  I will improve this to closer to 100%, but wanted to get
this in as it's still significantly better than before, which was 0%.
There are still a lot of tests in the code that are #if 0'd out
currently.  I still need to go through an enabled/delete those.

BUG=b:65025540
BRANCH=None
TEST=run COV=1 make and make sure passes without errors

Change-Id: If17bfc24fb7f65deb3d2286cc39cdc311d8a6d93
Signed-off-by: Shelley Chen <shchen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/830680
Reviewed-by: Randall Spangler <rspangler@chromium.org>
(cherry picked from commit e109b1e8d139ba3a9b7aeac2f09157c3b9a70824)
Reviewed-on: https://chromium-review.googlesource.com/861321