cgpt: show verbose details when primary ignored
A later change I authored has the tests check the GUID of the drive.
When the primary table is ignored the GUID from the secondary wasn't
being displayed either.  This change has the details of the secondary
table get displayed when the primary table is ignored.

BRANCH=none
BUG=None
TEST=Compiled.  The change was runtime tested as part of a larger change though
I didn't unit test this particular change after it was split out into a
separate commit.

Change-Id: I300511cf65c67f4888e08ab49cd72c7acf234507
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1173410
Reviewed-by: Julius Werner <jwerner@chromium.org>