futility: cmd_update: Access system firmware using external flashrom
To manipulate the firmware contents on device, we need to access the
flash chipset (usually via SPI) on system. The `host_flashrom` provides
a way to call external program "flashrom" for reading and writing
firmware. So the `update_firmware` can now load "system current
firmware" using flashrom.

Note in the future we may want to statically link the flashrom as
library so there won't be external dependency.

BUG=chromium:875551
TEST=make futil; futility update -i IMAGE
     tests/futility/run_test_scripts.sh $(pwd)/build/futility
BRANCH=None

Change-Id: I52f2d4fe4fe4dd660f762a5a75e3367820717e19
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1183650
Reviewed-by: Randall Spangler <rspangler@chromium.org>