crossystem: automate writable desc (and display type)
A bunch of the params have '(writable)' at the end of the description
to indicate it's a writable field.  However, it's not listed on every
field.  Rather than resync all of them, automate it.  Throw in the
type for good measure.

The old display:
  hwid          = LUMPY  # Hardware ID
  dev_boot_usb  = 1      # Enable developer mode boot from USB/SD (writable)
The new display:
  hwid          = LUMPY  # [RO/str] Hardware ID
  dev_boot_usb  = 1      # [RW/int] Enable developer mode boot from USB/SD

BUG=None
TEST=`crossystem` output looks better
BRANCH=None

Change-Id: I953cf5cb78b52edeece4215c3249b79b26d36f26
Reviewed-on: https://chromium-review.googlesource.com/1224652
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>