image_signing: drop support for <R16 recovery image signing
We changed the verity kernel command line form before R16 was released
and included backwards compat support in the scripts for it.  But all
the devices that were released for these old versions are EOL, and we
don't need to sign images that old anymore, so drop support.

BRANCH=None
BUG=chromium:891015
TEST=precq passes

Change-Id: I0e61c5d5cbeefb8ea0af955ead604a97fcb84bad
Reviewed-on: https://chromium-review.googlesource.com/1255344
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>