futility: updater: Refactor how arguments were configured
We are going to have more command line arguments that must be passed to
updater_setup_config, and it is better to manage so many variables in a
struct.

Also, revised the order or argument processing so that simple settings
are now processed first, then complicated ones or those with dependency.

BUG=chromium:875551
TEST=TEST=make futil; tests/futility/run_test_scripts.sh $(pwd)/build/futility
BRANCH=None

Change-Id: I03ac036d26e49cdf924c03d6e86a272ce89fc2aa
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1265575
Reviewed-by: Randall Spangler <rspangler@chromium.org>