futility: updater: Check EC/PD WP state again before updating
There are devices, especially during or after RMA, may have WP states
not synced; for example
 HW =      1
 SW (AP) = 0
 SW (EC) = 1

In this case, we can still update host firmware but not EC. This happens
more often on EC that needs an extra reboot to change WP states.

As a result, we do want to check real programmer again before updating
optional images.

BUG=chromium:902546
TEST=make futil; tests/futility/run_test_scripts.sh $(pwd)/build/futility
BRANCH=None

Change-Id: I9a526cde19a1ab3c41afecb4f7247bd941edc3f4
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1322295
Reviewed-by: Randall Spangler <rspangler@chromium.org>