Add console logging to make UI easier to navigate without a screen
Many people seem to have problems doing things in the detachable UI on
early bring-up devices that don't yet support display output. This patch
is supposed to alleviate that problem by adding more log output so that
people can see on the UART which menu option they have selected (with
actual text rather than just screen and menu item indices).

While we're at it, let's also dump the TAB output to the console so that
it's easier to retrieve that information without a display as well.

BRANCH=None
BUG=chromium:924446
TEST=None, because make runtests is broken and nobody seems to care.

Change-Id: If9350255a68821b7e232726ba56d001571d52cd4
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1436494
Reviewed-by: Randall Spangler <rspangler@chromium.org>