firmware/lib: commit nvram before running legacy
vb2_run_altfw() can jump to run a legacy payload, so this change
adds a call to vb2_nv_commit() to commit any pending changes
to nvram before making the jump.

The call to commit requires a vb2_context, so the majority of this
change is to plumb this context through various functions.

BUG=b:124358784
BRANCH=None
TEST=local compile.  Tested with a later change that helps confirm
that a pending nvram change is written before jumping.

Change-Id: Ib32980527aa07357d62dd695a6ff479e8c918cf8
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1504757
Reviewed-by: Julius Werner <jwerner@chromium.org>