image_signing: avoid using mount -o rw,remount
The comments said remount was avoided due to loop offset usage which,
while true, isn't the only reason.  We can't remount rw without first
calling enable_rw_mount to clear the invalid bits in the fs header.
Update the comment to reflect that and switch to our mount helper as
it will call that for us.

BUG=chromium:938958
TEST=precq passes & signing unittests pass
BRANCH=None

Change-Id: I1063bc84befebddc942a3dec05e8f33ea834db30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/1522089
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>