Make vboot -Wtype-limits compliant
-Wtype-limits is an additional warning we're trying to enable in
coreboot that catches common coding mistakes (e.g. checking whether an
unsigned variable is < 0). vboot almost works with this out of the box,
but there's one instance where we want such a check (because the
constant it's checking may change). This patch pragma's it out so that
we can still build with the new warning.

BRANCH=None
BUG=chromium:960270
TEST=make runtests

Change-Id: I678a5915c99451b7e0a2672efb5ae6c81ebfb027
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1598720
Commit-Ready: Joel Kitching <kitching@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>