SARIEN ONLY: do not wait for power button release
In order to be compatible with the current cr50 firmware this change
will make the user confirmation screen not wait for power button
release and instead return true when it is pressed.

Since it is no longer looking for power button release it also reads
and ignores the power button on recovery ui start, in case the power
button was held down too long when doing Refresh+Power.

BUG=b:129471321
BRANCH=sarien (ONLY)
TEST=successfully enter developer mode on sarien

Change-Id: I0b4d1e336b4fdfca102058543ca5d7d0dc055fff
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/1626930
Reviewed-by: Julius Werner <jwerner@chromium.org>