host: add host library for accessing chromeos-config
Host side library for accessing chromeos-config. Initially, this will
be used by futility to access the /firmware:image-name property during
a firmware update.

More background: go/mosys-firmware-name

(note: despite the name "mosys" in the design doc, this is an effort
to *not* rely on mosys during the firmware update)

BUG=chromium:1061192
BRANCH=none
TEST=provided unit tests

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ib8e5f8f836a93695e3b30731ae227501f37c4633
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2099449
Reviewed-by: Julius Werner <jwerner@chromium.org>