vboot/ui: Add refresh flag to support force refresh the screen in action
CL:2163088 is working on depthcharge side to support refreshing
everything in vboot side, but it is still work in progress.
In the meanwhile, we can have a flag to control the refreshing for
already support refreshing everytime screen.

This flag can be removed after all screens are supported refreshing
everytime in vboot.

BRANCH=none
BUG=b:156692539, b:156693348
TEST=emerge-hatch vboot_reference
TEST=verfied in later CL that the screen is refreshed when the action
callback set this flag.

Signed-off-by: Meng-Huan Yu <menghuan@chromium.org>
Change-Id: Ie76e8dce5186d0acfbc176a4cc8079c3df37fd38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2336239
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>